% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decon.R
\name{decon}
\alias{decon}
\title{Main function for deconvolving signal using 3D contacts}
\usage{
decon(y, D, x = NULL, df, gamma = 1)
}
\arguments{
\item{y}{vector of observed counts}

\item{D}{a symmetric matrix specifying the pairwise interactions. Should
be smoothed and normalized. Should be nxn where n is the length of y.}

\item{x}{(optional) vector of observed spatial positions. If not specified,
it is assumed that the counts y are evenly spaced.}

\item{df}{integer, sets the number of degrees of freedom to use for the
smoothing spline.}

\item{gamma}{(optional) weight for how much to penalize the off-diagonal
terms of D. default is 1, or no penalty.}
}
\value{
A list containing the
following output:
\itemize{ \item{\code{y}}{ a vector of the predicted deconvolved counts }
\item{\code{a}}{ a vector of the coefficient estimates}}.
}
\description{
This function fits a spatial beta-binomial model, estimates a
2D smoothed probability function, and stores information for plotting results.
Uses a likelihood ratio test to test for a significant spatial fit over
optional baseline covariates such as cell type.
}
\details{
Optimizes the log-likelihood using BFGS and returns the deconvolved signal.
}
