% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolve.R
\name{deconvolve,Dcon-method}
\alias{deconvolve,Dcon-method}
\title{Optimizes the log-likelihood using BFGS and returns the deconvolved signal.}
\usage{
\S4method{deconvolve}{Dcon}(object, df = 21, gamma = 0.5, pad = 10, add_center = 1)
}
\arguments{
\item{df}{integer, sets the number of degrees of freedom to use for the
smoothing spline.}

\item{gamma}{(optional) weight for how much to penalize the off-diagonal
terms of D. default is 1, or no penalty.}

\item{y}{vector of observed counts}

\item{D}{a symmetric matrix specifying the pairwise interactions. Should
be smoothed and normalized. Should be nxn where n is the length of y.}

\item{x}{(optional) vector of observed spatial positions. If not specified,
it is assumed that the counts y are evenly spaced.}
}
\value{
A list containing the
following output:
\itemize{ \item{\code{y}}{ a vector of the predicted deconvolved counts }
\item{\code{a}}{ a vector of the coefficient estimates}}.
}
\description{
Description
}
