% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_decon.R
\name{fit_decon}
\alias{fit_decon}
\title{Fit the 1D deconvolution model}
\usage{
fit_decon(reads, D, df = NULL, offset = NULL, hessian = F)
}
\arguments{
\item{reads}{a vector of counts}

\item{D}{a (normalized) matrix representing the DNA-DNA contacts. See
\code{normalize_hic} to normalize a matrix of counts. Must be symmetric with
number of rows and columns equal to the length of reads.}

\item{df}{the degrees of freedom for the Gaussian basis functions. The default
is the length of reads.}

\item{offset}{a vector of offset counts the same length as reads.}
}
\value{
a list containing \code{a}, the estimated coefficients, \code{B}, the
basis functions, \code{est} the estimated log-scale signal and \code{var} the variance
associated with the estimates.
}
\description{
Fits the 1D version of the deconvolution model, where a 1D
vector of counts gets convolved by a 2D matrix
}
