% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_pairwise_diff}
\alias{get_pairwise_diff}
\title{Get pairwise L1 and L2 norm differences from a list of matrices
and the indices at those matrices to calculate the differences on}
\usage{
get_pairwise_diff(matrix_list, idx1, idx2, label = "")
}
\description{
Get pairwise L1 and L2 norm differences from a list of matrices
and the indices at those matrices to calculate the differences on
}
