% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_anchors}
\alias{load_anchors}
\title{Load in anchors
If there are multiple anchor files, can decide whether to use all unique
entries or only take ones in both (intersection). Default is to take the
intersection.}
\usage{
load_anchors(..., coords, type = "fithichip", use = "intersection")
}
\arguments{
\item{...}{one or more file names to input.}

\item{type}{the type of the anchor file. Supports fithichip}

\item{use}{whether or not to use the intersection or union if there are
multiple input files}
}
\description{
Load in anchors
If there are multiple anchor files, can decide whether to use all unique
entries or only take ones in both (intersection). Default is to take the
intersection.
}
