% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_coords}
\alias{load_coords}
\title{Load in the hic-pro coordinates file
aka the _abs.bed file with the correct resolution bins}
\usage{
load_coords(filename, type = "hic-pro")
}
\description{
Load in the hic-pro coordinates file
aka the _abs.bed file with the correct resolution bins
}
