% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_hic.R
\name{normalize_hic}
\alias{normalize_hic}
\title{Normalize Hi-C or DNA-DNA contact matrix before deconvolution}
\usage{
normalize_hic(M, gamma = 1, threshold = 5, smooth = F, sigma = 1)
}
\arguments{
\item{M}{the matrix of raw counts}

\item{gamma}{the amount of penalty on the off-diagonal between 0-1. Default is
1 (no penalty)}

\item{threshold}{the minimum number of counts in a column.
If a column has less than this number, it does not participate in
deconvolution. Default is 5 (but usually should go higher)}

\item{smooth}{boolean indicating whether or not to smooth the contact
matrix. Smoothing is accomplished with \code{spatstat::blur}. The operations
are moderately time intensive and this probably needs to be sped up at
some point, so default is \code{FALSE}.}

\item{sigma}{if smooth is \code{TRUE}, the sd of the Gaussian kernel smooth. Default
is 1. May need to tune this if experimenting with different sized M.}
}
\value{
a list containing \code{a}, the estimated coefficients, \code{B}, the
basis functions, \code{est} the estimated log-scale signal and \code{var} the variance
associated with the estimates.
}
\description{
Column-normalizes and optionally smooths
}
