% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mat.R
\name{plot_mat}
\alias{plot_mat}
\title{Plot a matrix in Hi-C style}
\usage{
plot_mat(mat, max = NULL, style = "higlass", main = "", xlab = "", ylab = "")
}
\arguments{
\item{mat}{the matrix to plot; will use row and colnames for labeling}

\item{max}{the maximum value for plotting}

\item{style}{can be 'juicebox' or 'higlass' (default)}
}
\description{
Style can be Juicebox or higlass.
Huge matrices can take a long time to plot!!!
}
