% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{simulate_D}
\alias{simulate_D}
\title{Simulate a convolving matrix that resembles a typical DNA-DNA contact matrix
decay is the ratio between each consecutive diagonal
add_noise is the rate parameter for adding Poisson noise
if tads > 1, adds in equal-length tads by dividing len/tads}
\usage{
simulate_D(
  len,
  decay = 1/3,
  gamma = 1,
  add_noise = NULL,
  tads = 1,
  tad_strength = 0.2
)
}
\description{
Simulate a convolving matrix that resembles a typical DNA-DNA contact matrix
decay is the ratio between each consecutive diagonal
add_noise is the rate parameter for adding Poisson noise
if tads > 1, adds in equal-length tads by dividing len/tads
}
