% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{simulate_log_rate}
\alias{simulate_log_rate}
\title{Simulate the log(rate) using a linear combination of Gaussian basis functions
assuming that the coefficients are normally distributed with specified
mean and variance}
\usage{
simulate_log_rate(
  len,
  df,
  alpha_mean = -1,
  alpha_sd = 2,
  seed = 1,
  npeaks = NULL
)
}
\description{
Simulate the log(rate) using a linear combination of Gaussian basis functions
assuming that the coefficients are normally distributed with specified
mean and variance
}
