% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{simulate_y}
\alias{simulate_y}
\title{Simulate the Poisson-distributed outcome y}
\usage{
simulate_y(
  len,
  df,
  D = NULL,
  alpha_mean = -1,
  alpha_sd = 2,
  decay = 1/3,
  gamma = 1,
  npeaks = NULL
)
}
\description{
Simulate the Poisson-distributed outcome y
}
